% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Identifies outliers using the specified method.}
\usage{
detect_outliers(df, outlier_method = "IQR", drop_method = "obs", PCA = TRUE)
}
\arguments{
\item{df}{a matrix or data frame of numeric variables.}

\item{outlier_method}{a matrix or data frame of numeric variables.}

\item{drop_method}{a matrix or data frame of numeric variables.}

\item{method}{a matrix or data frame of numeric variables.}
}
\value{
A reduced dataframe.
}
\description{
`detect_outliers()` Identifies outliers using either 3 sigma or IQR. By
default, a PCA is estimated first and the outliers are identified in the
error term. This option can be ignored if PCA is set to FALSE.
}
\examples{
detect_outliers(df=df,outlier_method="IQR",drop_method="obs", PCA=TRUE)
}
