% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{distance_transformation}
\alias{distance_transformation}
\title{Distance Transformation}
\usage{
distance_transformation(W, transform = "Inv")
}
\arguments{
\item{W}{a matrix with the weights of the Network calculated with TOM.}

\item{transform}{Either Inv for inverse, or Log for -log10 transformation.}
}
\value{
A new Distance Matrix.
}
\description{
`distance_transformation()` calculates the distance using Inverse or
Logit transformations.
}
\examples{
distance_transformation(W,"Inv")
distance_transformation(W,"Log")
}
