% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{floyd_warshall}
\alias{floyd_warshall}
\title{Floyd Warshall}
\usage{
floyd_warshall(Distance)
}
\arguments{
\item{Distance}{a V*V Distance Matrix.}
}
\value{
A list with the minimum distance matrix, and the Next matrix with
the closest path.
}
\description{
`floyd_warshall()` calculates the minimum distance path between the nodes.
Outputs two matrices: min_Distance is the new Matrix containing the minimum
distance, and Next contains a matrix with the minimum distance neighbor
}
\examples{
floyd_warshall(Distance)
}
