% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{hc_fun}
\alias{hc_fun}
\title{Hierarchical cluster analysis using TOM Distance.}
\usage{
hc_fun(D, linkage = "ward.D")
}
\arguments{
\item{D}{the distance matrix of the Lower Level Network calculated with TOM.}

\item{linkage}{type of linkage for the cluster algorithm.}
}
\value{
A list with the new group index, the silhouette plot, and the best K.
}
\description{
`hc_fun()` does a hierarchical cluster analysis on the TOM distance matrix
D and chooses the best number of clusters K by maximizing the silhouette
values for each possible number of clusters.
}
\examples{
hc_fun(D)
}
