% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{key_paths}
\alias{key_paths}
\title{Key Paths}
\usage{
key_paths(paths)
}
\arguments{
\item{path}{a path resulting from get_path().}
}
\value{
a data frame with all the hops with the number of times it has been
used, and a plot.
}
\description{
`key_paths()` finds the key paths in a vector of paths by breaking them by
pairs and counting the amount of times each pair appears.
}
\examples{
split_path(path)
}
