% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{mat_partial_cor_fun}
\alias{mat_partial_cor_fun}
\title{Estimates the partial correlation matrix of a data frame DF.}
\usage{
mat_partial_cor_fun(DF, Z, cor_method)
}
\arguments{
\item{DF}{the data frame with X and Y variables.}

\item{Z}{a data frame with the conditioning variables.}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance) is to be computed.
One of pearson (default), spearman or kendall.}
}
\value{
A data frame with the estimated partial correlation and the P. Value.
}
\description{
`mat_partial_cor_fun()` estimates the partial correlation matrix of DF
adjusting by Z variables.
}
\examples{
mat_partial_cor_fun(Z, DF, "pearson")
}
\seealso{
[partial_cor_fun].
}
