% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{multi_network}
\alias{multi_network}
\title{Multi Layer Network Analysis}
\usage{
multi_network(
  DF,
  Groups = NULL,
  Z = NULL,
  cor_method = "pearson",
  power = 6,
  TOMType = "unsigned",
  TOMDenom = "min",
  ClGroup = "G1",
  ClType = "HC",
  writeCSV = FALSE
)
}
\arguments{
\item{DF}{a data frame used to estimate the networks.}

\item{Groups}{a vector with the corresponding group of each variable. If
not specified, the groups will be optimized with a clustering algorithm.}

\item{Z}{a data frame for adjusting the partial correlations. If not
specified, we will use correlation instead of partial correlation.}

\item{cor_method}{a character variable that specifies the correlation
method. Default is "pearson".}

\item{power}{the power function for the soft threshold. Default is 6.}

\item{TOMType}{type of TOM function. Default is unsigned.}

\item{TOMDenom}{type of function in the denominator. Default is min.}

\item{ClGroup}{a specific group to do the clustering analysis.}

\item{ClType}{one of "HC" (Hierarchical Clustering), "PAM"
(K Medoids clustering) or "NOCL" (No Cluster). Default if "HC".}

\item{writeCSV}{set TRUE if you want to save your output into CSV files. The
default is FALSE.}
}
\value{
A list with the Upper and Lower Level Networks, the correlation
matrix, a data frame with the groups index, and the best K and the
silhouette plot if any clustering algorithm was used.
}
\description{
`multi_network()` This code applies the Multivariate Network Analysis
methodology used in Guan, Cheng, Koo.
}
\details{
First, a lower-level network based on correlation matrix is constructed to
represent interaction patterns between each pair of variables. Soft
thresholding (power function and Topological Overlap Measure) was applied to
the matrix to emphasize strong connections and remove spurious connections.

Then, to improve visualization of extensive network structure from the high
dimensional data, biologically predefined domains (i.e., blood lipid,
cytokines, etc.) and hierarchical cluster analysis is used to define modules
of potentially highly interconnected variables. The resulted upper-level
network represents average connectivity between modules.
}
\examples{
multi_network(DF,Groups,Z,ClGroup="Group 1",ClType="HC")
multi_network(DF,ClType="HC")
multi_network(DF,Groups,ClType="NOCL")
}
