% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{NA_replace}
\alias{NA_replace}
\title{Fills or drops NA values from a DF.}
\usage{
NA_replace(DF, Z = NULL, replace_method = "mean")
}
\arguments{
\item{DF}{a matrix or data frame with some NA.}

\item{replace_method}{replacement method for NA values: either drop or mean.}
}
\value{
A data frame without NA.
}
\description{
`NA_replace()` Fills or drops NA values from a DF.
}
\examples{
NA_replace(DF, "mean")
NA_replace(DF, Z, "mean")
}
