% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{outlier_detection_iqr}
\alias{outlier_detection_iqr}
\title{Identifies outliers using the specified method.}
\usage{
outlier_detection_iqr(x)
}
\arguments{
\item{x}{a matrix or data frame of numeric variables.}
}
\value{
A Boolean vector indicating the outliers.
}
\description{
`outlier_detection_iqr()` Identifies outliers which are 1.5 IQR higher than
the mean.
}
\examples{
outlier_detection_iqr(X)
}
