% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{partial_cor_fun}
\alias{partial_cor_fun}
\title{Estimates the partial correlation of X and Y given Z.}
\usage{
partial_cor_fun(x_varname, y_varname, Z, DF, cor_method)
}
\arguments{
\item{x_varname}{a character string indicating the name of X variable.}

\item{y_varname}{a character string indicating the name of Y variable.}

\item{Z}{a data frame with the conditioning variables.}

\item{DF}{the data frame with X and Y variables.}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance) is to be computed.
One of pearson (default), spearman or kendall.}
}
\value{
A data frame with the estimated partial correlation and the P. Value.
}
\description{
`partial_cor_fun()` look for variables "x_varname"and "y_varname" in DF and
estimates the partial correlation adjusting by Z.
}
\examples{
partial_cor_fun("x_varname", "y_varname", Z, DF, "pearson")
}
