% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{path_finder}
\alias{path_finder}
\title{Path Finder}
\usage{
path_finder(Next)
}
\arguments{
\item{Next}{a Matrix with the next closest neighbor between two nodes.}
}
\value{
a Matrix with the best paths for each pair of nodes.
}
\description{
`path_finder()` gets the best path between each pair of nodes.
}
\examples{
path_finder(Next)
}
