% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{pca_errors}
\alias{pca_errors}
\title{Fits a PCA and estimates the errors using 95% of the explained variance.}
\usage{
pca_errors(X)
}
\arguments{
\item{X}{a matrix or data frame of numeric variables.}
}
\value{
A reduced dataframe.
}
\description{
`pca_errors()` Fits a PCA and estimates the errors using 90% of the
explained variance.
}
\examples{
pca_errors(X)
}
