% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{replace_with_mean}
\alias{replace_with_mean}
\title{Fills NA values with the mean.}
\usage{
replace_with_mean(x)
}
\arguments{
\item{x}{a numeric vector with some NA.}
}
\value{
A numeric vector without NA.
}
\description{
`replace_with_mean()` look for variables "x_varname"and "y_varname" in DF and
estimates the partial correlation adjusting by Z.
}
\examples{
replace_with_mean(x)
}
