% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{split_path}
\alias{split_path}
\title{Split Path}
\usage{
split_path(path)
}
\arguments{
\item{path}{a path resulting from get_path().}
}
\value{
a string with the hops, separated by a comma.
}
\description{
`split_path()` split path into pairs of nodes.
}
\examples{
split_path(path)
}
