% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{uln_fun}
\alias{uln_fun}
\title{Upper Level Network.}
\usage{
uln_fun(W, groupsID)
}
\arguments{
\item{W}{the Lower Level Network calculated with TOM.}

\item{groupsID}{data frame with the groups information.}
}
\value{
A matrix with the weights W of the Upper Level NetworK.
}
\description{
`uln_fun()` calculates the Upper Level Network using the mean w_{ij} values
of all the units in the groups.
}
\examples{
uln_fun(groupsID,W)
}
