% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks Functions Final.R
\name{Wilcoxon_paired_test}
\alias{Wilcoxon_paired_test}
\title{Wilcoxon Paired Test}
\usage{
Wilcoxon_paired_test(X, Y, paired = T)
}
\arguments{
\item{X}{a Bootstrapped Network Matrix.}

\item{Y}{a Bootstrapped Network Matrix.}

\item{paired}{either TRUE of FALSE for Wilcoxon Paired Test.}
}
\value{
A list with the output of the wilcoxon test for each weight, and a
plot with the confidence interval and the median.
}
\description{
`Wilcoxon_paired_test()` performs a Wilcoxon test on the median of the
bootstrapped distributions of the weights, and compares the Networks of
two groups.
}
\examples{
split_path(path)
}
